+++ db/migrate/20240115183623_create_comment_replies.rb
@@ -1,8 +1,8 @@
 class CreateCommentReplies < ActiveRecord::Migration[6.1]
   def change
     create_table :comment_replies do |t|
-      t.integer :user_id, null: false
-      t.integer :comment_id, null: false
+      t.bigint :user_id, null: false
+      t.bigint :comment_id, null: false
 
       t.timestamps
     end
