+++ db/migrate/20240115183651_create_comments.rb
@@ -1,7 +1,7 @@
 class CreateComments < ActiveRecord::Migration[6.1]
   def change
     create_table :comments do |t|
-      t.integer :user_id, null: false
+      t.bigint :user_id, null: false
       t.text :posted_text, null: false
       t.text :reply, null: false
