+++ db/migrate/20240315182346_create_reports.rb
@@ -1,8 +1,8 @@
 class CreateReports < ActiveRecord::Migration[6.1]
   def change
     create_table :reports do |t|
-      t.integer :reporter_id
-      t.integer :reported_id
+      t.bigint :reporter_id
+      t.bigint :reported_id
       t.text :reason
 
       t.timestamps
